move new hp styling to the emergency hp page
