+++ frontend/lib/emergency-hp-action.tsx
@@ -27,34 +27,78 @@ import { performHardOrSoftRedirect } from './pages/login-page';
 import EMERGENCY_HPA_ISSUE_LIST from '../../common-data/emergency-hpa-issue-list.json';
 import { DjangoChoices } from './common-data';
 import { getIssueChoiceLabels, IssueChoice } from '../../common-data/issue-choices';
+import { MoratoriumWarning } from './covid-banners';
+import { StaticImage } from './static-image';
 
 const EMERGENCY_HPA_ISSUE_SET = new Set(EMERGENCY_HPA_ISSUE_LIST);
 
+const HP_ICON = "frontend/img/hp-action.svg";
+
 const onboardingForHPActionRoute = () => getSignupIntentOnboardingInfo(OnboardingInfoSignupIntent.EHP).onboarding.latestStep;
 
-const Disclaimer: React.FC<{}> = () => (
-  <div className="notification is-warning">
-    <p>Due to the COVID-19 pandemic, Housing Courts in New York City are only accepting cases for the following:</p>
-    <ul>
-      <li>Repairs for heat</li>
-      <li>Repairs for hot water</li>
-    </ul>
-  </div>
-);
+// HP Cases currently being accepted in Housing Court amidst COVID-19 crisis
+const acceptedCases = [
+  "no heat", 
+  "no hot water", 
+  "no gas", 
+  "mold", 
+  "lead-based paint", 
+  "no working toilet", 
+  "vacate order issued"
+];
+
+function Disclaimer(): JSX.Element {
+  const numCases = acceptedCases.length;
+  const generateCaseList = (start: number, end: number) => 
+    acceptedCases.map((caseType, i) => <li key={i}> {caseType} </li>).slice(start, end);
+  return (
+    <div className="notification is-warning">
+      <p>Due to the covid-19 pandemic, Housing Courts in New York City are only accepting cases for the following:</p>
+      <div className="is-hidden-tablet">
+        {generateCaseList(0,numCases)}
+      </div>
+      <div className="columns is-mobile is-hidden-mobile">
+        <div className="column is-one-third">
+          {generateCaseList(0,Math.round(numCases / 2))}
+        </div>
+        <div className="column">
+          {generateCaseList(Math.round(numCases / 2), numCases)}
+        </div>
+      </div>
+    </div>
+  );
+}
 
 function EmergencyHPActionSplash(): JSX.Element {
-  const {efnycOrigin} = useContext(AppContext).server;
-
   return (
-    <Page title="Sue your landlord for Repairs through an Emergency HP Action proceeding" withHeading="big" className="content">
-      <Disclaimer />
-      <p>Welcome to JustFix.nyc! This website will guide you through the process of starting an <strong>Emergency HP Action</strong> proceeding.</p>
-      <p>An <strong>Emergency HP Action</strong> is a legal case you can bring against your landlord for failing to make repairs and not providing essential services.</p>
-      <p><em>This service is free and secure.</em></p>
-      <p>Are you facing an eviction? It is important for you to get help ASAP. Visit <a href={efnycOrigin} target="_blank" rel="noreferrer noopener">EFNYC</a> to see if you qualify for a free lawyer.</p>
-      <GetStartedButton to={onboardingForHPActionRoute()} intent={OnboardingInfoSignupIntent.EHP} pageType="splash">
-        Start my case
-      </GetStartedButton>
+    <Page title="Sue your landlord for Repairs and/or Harassment through an HP Action proceeding">
+        <section className="hero is-light">
+          <div className="hero-body">
+            <div className="content has-text-centered">
+              <div className="is-inline-block jf-hp-icon">
+                <StaticImage ratio="is-square" src={HP_ICON} alt="" />
+              </div>
+              <h1 className="title is-spaced">
+                Sue your landlord for Repairs and/or Harassment through an HP Action proceeding
+              </h1>
+            </div>
+            <div className="columns is-centered">
+              <div className="column is-four-fifths">
+                <div className="content">
+                  <p className="subtitle">
+                    An HP Action is a legal case you can bring against your landlord for failing to make repairs, not providing essential services, or harassing you.
+                    This service is free, secure, and confidential.
+                  </p>
+                  <Disclaimer />
+                  <GetStartedButton to={onboardingForHPActionRoute()} intent={OnboardingInfoSignupIntent.HP} pageType="splash">
+                    Start my case
+                  </GetStartedButton>
+                </div>
+              </div>
+            </div>
+            <MoratoriumWarning />
+          </div>
+        </section>
     </Page>
   );
 }
@@ -65,22 +109,21 @@ const EmergencyHPActionWelcome = () => {
 
   return (
     <Page title={title} withHeading="big" className="content">
-      <Disclaimer />
+      <Disclaimer/>
       <p>
-        An <strong>Emergency HP (Housing Part) Action</strong> is a legal case you can bring against your landlord for failing to make repairs, and not providing essential services. Here is how it works:
+        An <strong>HP (Housing Part) Action</strong> is a legal case you can bring against your landlord for failing to make repairs, not providing essential services, or harassing you. Here is how it works:
       </p>
-      <ol className="has-text-left">
-        <li>You answer a few questions here about your housing situation and we email the forms you need to start your case to your email and your borough's Housing Court.</li>
-        <li>You will be assigned a lawyer who will help you throughout your case.</li>
-        <li>An inspector from Housing and Preservation and Development (HPD) will come to your house to verify the issue(s). Your lawyer will help you arrange a time that is convenient for you and give you the details you will need.</li>
-        <li>The court hearing will happen through a video call so that you do not have to go to the courthouse in-person. Your lawyer will give you all of the details and will guide you every step of the way.</li>
-      </ol>
-      <div className="notification is-warning">
-        <p>Due to the COVID-19 pandemic, Housing Courts in New York City will be conducting hearings via video conferencing. Tenants will not be required to go to Housing Court in person.</p>
-      </div>
-      <GetStartedButton to={Routes.locale.ehp.sue} intent={OnboardingInfoSignupIntent.EHP} pageType="welcome">
+      <BigList listClassName="is-light">
+        <li>You answer a few questions here about your housing situation and we email the forms you need to start your case to your email and your Borough’s Housing Court.</li>
+        <li>You will be assigned a Lawyer that will help you throughout your case.</li>
+        <li>An inspector from the Department of Housing and Preservation (HPD) will come to your house to verify the issue(s). Your Lawyer will help you arrange a time that is convenient for you and give you the details you will need.</li>
+        <li>The court hearing will happen through a video-call so that <strong>you do not have to go to the Courthouse in-person</strong>. Your Lawyer will give you all of the details and will guide you each step of the way.</li>
+      </BigList>
+        <br />
+      <GetStartedButton to={Routes.locale.hp.sue} intent={OnboardingInfoSignupIntent.HP} pageType="welcome">
         Get started
       </GetStartedButton>
+      <MoratoriumWarning />
     </Page>
   );
 };
